Make domain services do something rather than return something
Prior to this change, the domain service just returned an event rather
than "doing" something.

This change modifies that by passing the event list into the operations
and having them append events themselves. This means the operations
appear to be "doing" something, rather than just returning the event
that someone else needs to add to the committer.